# health-id-system/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Research API for the HEALTH_ID remote examination module
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /results:
    get:
      operationId: listResults
      tags: [calculations]
      summary: List HEALTH_ID results
      parameters:
        - $ref: "#/components/parameters/Limit"
        - in: query
          name: category
          schema:
            type: string
            enum: [green, yellow, red]
        - in: query
          name: worker_pseudonym
          schema:
            type: string
      responses:
        "200":
          description: Result list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResultList"
  /results/{resultId}:
    get:
      operationId: getResult
      tags: [calculations]
      summary: Get an explainable result
      parameters:
        - $ref: "#/components/parameters/ResultId"
      responses:
        "200":
          description: Result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResult"
        "404":
          $ref: "#/components/responses/NotFound"
  /calculate:
    post:
      operationId: calculateHealthId
      tags: [calculations]
      summary: Calculate HEALTH_ID for a research event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CalculationInput"
      responses:
        "200":
          description: Calculation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthResult"
  /dynamics/{workerPseudonym}:
    get:
      operationId: getWorkerDynamics
      tags: [calculations]
      summary: Get worker index dynamics
      parameters:
        - in: path
          name: workerPseudonym
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Dynamics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dynamics"
  /reviews:
    get:
      operationId: listReviews
      tags: [review]
      summary: List human review cases
      responses:
        "200":
          description: Review queue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReviewList"
    post:
      operationId: createReview
      tags: [review]
      summary: Create a human review case
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewInput"
      responses:
        "201":
          description: Review case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
  /reviews/{reviewId}:
    patch:
      operationId: updateReview
      tags: [review]
      summary: Resolve a human review case
      parameters:
        - $ref: "#/components/parameters/ReviewId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReviewUpdate"
      responses:
        "200":
          description: Updated review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
...
  /verify-identity:
    post:
      operationId: verifyIdentity
      tags: [verification]
      summary: Run sandbox identity verification
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerificationInput"
      responses:
        "200":
          description: Verification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"
  /verify/{sessionId}:
    get:
      operationId: getVerification
      tags: [verification]
      summary: Get verification session
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          description: Verification session
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerificationResult"
components:
  parameters:
    Limit:
      in: query
      name: limit
      schema:
        type: integer
        minimum: 1
        maximum: 200
        default: 50
    ResultId:
      in: path
      name: resultId
      required: true
      schema: { type: string }
    ReviewId:
      in: path
      name: reviewId
      required: true
      schema: { type: string }
    SessionId:
      in: path
      name: sessionId
      required: true
      schema: { type: string }
  responses:
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error: { type: string }
...
    CalculationInput:
      type: object
      required: [event_id, worker_pseudonym, measurements]
      properties:
        event_id: { type: string }
        worker_pseudonym: { type: string }
        measured_at: { type: string }
        measurements:
          $ref: "#/components/schemas/Measurements"
        model_version:
          type: string
          default: health_id_v1.0.0
    Measurements:
      type: object
      required: [heart_rate, systolic, diastolic, temperature, spo2, alcohol_test, adequacy_score, pupil_reaction, examination_regularity, missed_examinations]
      properties:
        heart_rate: { type: number }
        systolic: { type: number }
        diastolic: { type: number }
        temperature: { type: number }
        spo2: { type: number }
        alcohol_test: { type: number }
        adequacy_score: { type: number }
        speech_coherence: { type: ["number", "null"] }
        pupil_reaction: { type: number }
        examination_regularity: { type: number }
        missed_examinations: { type: number }
    HealthResult:
      type: object
      required: [result_id, event_id, worker_pseudonym, timestamp, health_id, components, completeness, uncertainty, state_flags, explanation, evidence, audit]
      properties:
        result_id: { type: string }
        event_id: { type: string }
        worker_pseudonym: { type: string }
        timestamp: { type: string }
        health_id:
          $ref: "#/components/schemas/HealthScore"
        components:
          type: array
          items:
            $ref: "#/components/schemas/ComponentScore"
        completeness:
          $ref: "#/components/schemas/Completeness"
        uncertainty:
          $ref: "#/components/schemas/Uncertainty"
        state_flags:
          type: array
          items:
            $ref: "#/components/schemas/StateFlag"
        explanation: { type: string }
        evidence:
          $ref: "#/components/schemas/Evidence"
        audit:
          $ref: "#/components/schemas/CalculationAudit"
    ResultList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/HealthResult"
        total: { type: integer }
    HealthScore:
      type: object
      required: [value, category, model_version, config_snapshot_hash, disclaimer]
      properties:
        value: { type: number }
        category:
          type: string
          enum: [green, yellow, red]
        model_version: { type: string }
        config_snapshot_hash: { type: string }
        disclaimer: { type: string }
    ComponentScore:
      type: object
      required: [key, label, value, weight, contribution, contribution_pct, features]
      properties:
        key: { type: string }
        label: { type: string }
        value: { type: number }
        weight: { type: number }
        contribution: { type: number }
        contribution_pct: { type: number }
        features:
          type: array
          items:
            $ref: "#/components/schemas/FeatureEvidence"
    FeatureEvidence:
      type: object
      required: [name, label, raw_value, unit, source, normalized_score, normal_range, contribution_weight, contribution, quality, deviation]
      properties:
        name: { type: string }
        label: { type: string }
        raw_value: { type: ["number", "null"] }
        unit: { type: string }
        source:
          type: string
          enum: [measured, derived, context]
        normalized_score: { type: number }
        normal_range:
          type: array
          items: { type: number }
        contribution_weight: { type: number }
        contribution: { type: number }
        quality: { type: string }
        deviation: { type: string }
    Completeness:
      type: object
      required: [overall, required_features_present, required_features_total, missing]
      properties:
        overall: { type: number }
        required_features_present: { type: integer }
        required_features_total: { type: integer }
        missing:
          type: array
          items: { type: string }
    Uncertainty:
      type: object
      required: [overall, sources]
      properties:
        overall: { type: number }
        sources:
          type: array
          items:
            type: object
            required: [type, impact]
            properties:
              type: { type: string }
              feature: { type: ["string", "null"] }
              impact: { type: number }
    StateFlag:
      type: object
      required: [flag, active, label]
      properties:
        flag: { type: string }
        active: { type: boolean }
        label: { type: string }
        detail: { type: ["string", "null"] }
    Evidence:
      type: object
      required: [input_ids, formulas_applied, baseline_used, quality_summary]
      properties:
        input_ids:
          type: array
          items: { type: string }
        formulas_applied:
          type: array
          items: { type: string }
        baseline_used:
          $ref: "#/components/schemas/BaselineSummary"
        quality_summary: { type: string }
    BaselineSummary:
      type: object
      required: [available, n_historical_points, window_days, excluded_event]
      properties:
        available: { type: boolean }
        n_historical_points: { type: integer }
        window_days: { type: integer }
        excluded_event: { type: string }
    CalculationAudit:
      type: object
      required: [calculation_timestamp, duration_ms, reproducible, audit_chain_hash]
      properties:
        calculation_timestamp: { type: string }
        duration_ms: { type: integer }
        reproducible: { type: boolean }
        audit_chain_hash: { type: string }
    Review:
      type: object
      required: [review_id, session_id, event_id, status, trigger_reason, created_at, assigned_to]
      properties:
        review_id: { type: string }
        session_id: { type: string }
        event_id: { type: string }
        status:
          type: string
          enum: [pending, confirmed, rejected]
        trigger_reason: { type: string }
        created_at: { type: string }
        resolved_at: { type: ["string", "null"] }
        assigned_to: { type: string }
        reviewer_comment: { type: string }
        match_score: { type: ["number", "null"] }
    ReviewList:
      type: object
      required: [items, total]
      properties:
        items:
          type: array
          items: { $ref: "#/components/schemas/Review" }
        total: { type: integer }
    ReviewInput:
      type: object
      required: [session_id, event_id, trigger_reason]
      properties:
        session_id: { type: string }
        event_id: { type: string }
        trigger_reason: { type: string }
        match_score: { type: number }
    ReviewUpdate:
      type: object
      required: [status, reviewer_comment]
      properties:
        status:
          type: string
          enum: [confirmed, rejected]
        reviewer_comment: { type: string, minLength: 1 }
...
    Dynamics:
      type: object
      required: [worker_pseudonym, points, baseline]
      properties:
        worker_pseudonym: { type: string }
        points:
          type: array
          items:
            type: object
            required: [date, value, category]
            properties:
              date: { type: string }
              value: { type: number }
              category: { type: string }
        baseline:
          $ref: "#/components/schemas/BaselineSummary"
    VerificationInput:
      type: object
      required: [worker_pseudonym, match_score, liveness_score, quality_pass]
      properties:
        worker_pseudonym: { type: string }
        match_score: { type: number, minimum: 0, maximum: 1 }
        liveness_score: { type: number, minimum: 0, maximum: 1 }
        quality_pass: { type: boolean }
        active_liveness: { type: boolean, default: true }
    VerificationResult:
      type: object
      required: [session_id, route, match_score, liveness_score, quality_pass, model_version, disclaimer]
      properties:
        session_id: { type: string }
        route:
          type: string
          enum: [verified, manual_review, not_verified]
        match_score: { type: number }
        liveness_score: { type: number }
        quality_pass: { type: boolean }
        model_version: { type: string }
        created_at: { type: string }
        disclaimer: { type: string }