# health-id-system/lib/api-zod/src/generated/types/calculationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Research API for the HEALTH_ID remote examination module
 * OpenAPI spec version: 0.1.0
 */
import type { Measurements } from './measurements';

export interface CalculationInput {
  event_id: string;
  worker_pseudonym: string;
  measured_at?: string;
  measurements: Measurements;
  model_version?: string;
}