# health-id-system/lib/api-zod/src/generated/types/error.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Research API for the HEALTH_ID remote examination module
 * OpenAPI spec version: 0.1.0
 */

export interface Error {
  error: string;
}