# health-id-system/lib/api-zod/src/generated/types/evidence.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Research API for the HEALTH_ID remote examination module
 * OpenAPI spec version: 0.1.0
 */
import type { BaselineSummary } from './baselineSummary';

export interface Evidence {
  input_ids: string[];
  formulas_applied: string[];
  baseline_used: BaselineSummary;
  quality_summary: string;
}