# health-id-system/lib/api-zod/src/generated/types/measurements.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Research API for the HEALTH_ID remote examination module
 * OpenAPI spec version: 0.1.0
 */

export interface Measurements {
  heart_rate: number;
  systolic: number;
  diastolic: number;
  temperature: number;
  spo2: number;
  alcohol_test: number;
  adequacy_score: number;
  /** @nullable */
  speech_coherence?: number | null;
  pupil_reaction: number;
  examination_regularity: number;
  missed_examinations: number;
}